refactor: setup function
