+++ package.json
@@ -24,6 +24,7 @@
     "test:bridge:webpack": "TEST_BRIDGE=1 yarn test:presets",
     "test:bridge:vite": "TEST_BRIDGE_VITE=1 TEST_BRIDGE=1 yarn test:presets",
     "test:unit": "vitest packages",
+    "test:utils": "vitest test/examples",
     "version": "yarn && git add yarn.lock"
   },
   "resolutions": {
