+++ packages/test-utils/src/setup/vitest.ts
@@ -1,0 +1,9 @@
+import type { TestHooks } from '../types'
+
+export default async function setupVitest (hooks: TestHooks) {
+  const vitest = await import('vitest')
+  vitest.test('setup', hooks.setup, 60000)
+  vitest.beforeEach(hooks.beforeEach)
+  vitest.afterEach(hooks.afterEach)
+  vitest.afterAll(hooks.afterAll)
+}
